Updated HostToHostIntentCompiler: 	Parsing annotation values to integer in addConstraintValues method Fixed AdvancedAnnotationConstraint: 	Getting link directly from link store when checked for validity.
	modified:   core/api/src/main/java/org/onosproject/net/intent/constraint/AdvancedAnnotationConstraint.java
	modified:   core/net/src/main/java/org/onosproject/net/intent/impl/compiler/HostToHostIntentCompiler.java

+++ core/net/src/main/java/org/onosproject/net/intent/impl/compiler/HostToHostIntentCompiler.java
@@ -239,13 +239,16 @@ public class HostToHostIntentCompiler
 
                     if (advConstraint.isUpperLimit()) {
                         // is upper limit -> increase annotation value
-                        return String.valueOf(Double.valueOf(linkAnnotationValue) + advConstraint.threshold());
+                        return String.valueOf(((Double) (Double.valueOf(linkAnnotationValue)
+                                + advConstraint.threshold())).intValue());
                     } else {
-                        return String.valueOf(Double.valueOf(linkAnnotationValue) - advConstraint.threshold());
+                        return String.valueOf(((Double) (Double.valueOf(linkAnnotationValue)
+                                - advConstraint.threshold())).intValue());
                     }
                 } else {
                     // AnnotationConstraint has always an upper limit -> increase annotation value
-                    return String.valueOf(Double.valueOf(linkAnnotationValue) + annotationConstraint.threshold());
+                    return String.valueOf(((Double) (Double.valueOf(linkAnnotationValue)
+                            + annotationConstraint.threshold())).intValue());
                 }
 
             }
